fix optical flow

+++ optical_flow.py
@@ -2,9 +2,9 @@ from video_model import VideoModel
 import cv2
 import numpy as np
 import torch
-
-FIRE_LOWER_HSV = np.array([0, 100, 100])
-FIRE_UPPER_HSV = np.array([10, 255, 255])
+DEBUG = False
+FIRE_LOWER_RGB = np.array([150,  50,   0])   # R, G, B
+FIRE_UPPER_RGB = np.array([255, 255, 150])
 def draw_flow_on_white(flow, shape, step=16):
     """
     Draws optical flow vectors as arrows on a white background.
@@ -48,6 +48,10 @@ class FarnebackOpticalFlowModel(VideoModel):
         
         prev_frame_np = frames[-2].permute(1,2,0).numpy()
         target_frame_np = frames[-1].permute(1,2,0).numpy()
+
+        # # print max and min values of the frames
+        # print(f"prev_frame_np max: {np.max(prev_frame_np)}, min: {np.min(prev_frame_np)}")
+        # print(f"target_frame_np max: {np.max(target_frame_np)}, min: {np.min(target_frame_np)}")
         
         # Convert the last two frames to grayscale, and convert (C, H, W) to (H, W, C)
         prev_frame = cv2.cvtColor(prev_frame_np, cv2.COLOR_RGB2GRAY)
@@ -57,11 +61,9 @@ class FarnebackOpticalFlowModel(VideoModel):
         flow = cv2.calcOpticalFlowFarneback(prev_frame, pred_frame, None, 0.5, 3, 15, 3, 5, 1.2, 0)
         # Calculate the magnitude of the flow vectors
         magnitude, angle = cv2.cartToPolar(flow[..., 0], flow[..., 1])
-        # Create a mask for the fire color range (orange-red) in HSV
-        # Convert the last frame to HSV color space
-        hsv_frame = cv2.cvtColor(target_frame_np, cv2.COLOR_BGR2HSV)
-        # Create a mask for the fire color range
-        fire_mask = cv2.inRange(hsv_frame, FIRE_LOWER_HSV, FIRE_UPPER_HSV)
+        # Create a mask for the fire color range (orange-red)
+        # we mask previous frame since flow[i,j] = (dx, dy) represents the flow from the previous frame to the current frame, so we want to observe flow in pixels where there was previously fire-like color
+        fire_mask = cv2.inRange(prev_frame_np, FIRE_LOWER_RGB, FIRE_UPPER_RGB)
         # Calculate the sum of the flow magnitudes in the fire regions
         flow_magnitude_in_fire = np.sum(magnitude[fire_mask > 0])
         # Calculate the total flow magnitude and magnitude of the flow in the fire regions
@@ -86,12 +88,12 @@ class FarnebackOpticalFlowModel(VideoModel):
         flow_rgb = cv2.cvtColor(hsv, cv2.COLOR_HSV2RGB)
 
 
-
-        cv2.imshow("Flow Field", draw_flow_on_white(flow, target_frame_np.shape[:2]))
-        cv2.imshow("Original Frame", cv2.cvtColor(target_frame_np, cv2.COLOR_BGR2RGB))
-        cv2.imshow("Fire Mask", fire_mask)
-        cv2.waitKey(0)
-        cv2.destroyAllWindows()
+        if DEBUG:
+            cv2.imshow("Flow Field", draw_flow_on_white(flow, target_frame_np.shape[:2]))
+            cv2.imshow("Original Frame", cv2.cvtColor(target_frame_np, cv2.COLOR_BGR2RGB))
+            cv2.imshow("Fire Mask", fire_mask)
+            cv2.waitKey(0)
+            cv2.destroyAllWindows()
 
 
         return fire_prob
